Summarization/requirements.txt
datasets>=2.19.0
transformers>=4.41.0
torch>=2.2.0
rouge-score>=0.1.2
sentence-transformers>=2.7.0
psutil>=5.9.8
pandas>=2.2.2
tqdm>=4.66.4
google-genai>=1.30.0
